Convert mailmanager into an object. This makes it easy to exclude from test coverage
