+++ feature/diary-auth/src/commonMain/kotlin/com/foreverrafs/superdiary/auth/MailManager.kt
@@ -1,0 +1,5 @@
+package com.foreverrafs.superdiary.auth
+
+expect object MailManager {
+    fun openMail()
+}
