--- feature/diary-auth/src/commonMain/kotlin/com/foreverrafs/superdiary/auth/OpenMail.kt
@@ -1,3 +1,0 @@
-package com.foreverrafs.superdiary.auth
-
-expect fun openMail()
