+++ feature/diary-auth/src/iosMain/kotlin/com/foreverrafs/superdiary/auth/MailManager.kt
@@ -1,0 +1,18 @@
+package com.foreverrafs.superdiary.auth
+
+import platform.Foundation.NSURL
+import platform.UIKit.UIApplication
+
+actual object MailManager {
+    actual fun openMail() {
+        NSURL.URLWithString("message://")?.let {
+            if (UIApplication.sharedApplication.canOpenURL(url = it)) {
+                UIApplication.sharedApplication.openURL(
+                    url = it,
+                    options = emptyMap<Any?, Any?>(),
+                    completionHandler = {},
+                )
+            }
+        }
+    }
+}
