--- feature/diary-auth/src/jvmMain/kotlin/com/foreverrafs/superdiary/auth/OpenMail.kt
@@ -1,14 +1,0 @@
-package com.foreverrafs.superdiary.auth
-
-import java.awt.Desktop
-import java.io.IOException
-
-actual fun openMail() {
-    try {
-        if (Desktop.isDesktopSupported() && Desktop.getDesktop().isSupported(Desktop.Action.MAIL)) {
-            Desktop.getDesktop().mail()
-        }
-    } catch (e: IOException) {
-        e.printStackTrace()
-    }
-}
