+++ settings.gradle.kts
@@ -98,7 +98,7 @@ kover {
         "*utils.FileSystem_androidKt",
         "*generated.resources.*",
         "*components.*",
-        "*OpenMail.*",
+        "*MailManager",
     )
 
     reports {
